feat(backend): 포인트 불변 원장 기반 구축
Flyway 마이그레이션과 포인트 엔티티·저장소 추가
기존 잔액 이관 및 프로필 잔액 조회 전환

+++ backend/src/main/java/com/dailyatelier/dailyatelier/dto/UserProfileDto.java
@@ -11,7 +11,7 @@ public class UserProfileDto {
     private String phoneNumber;
     private String email;
     private int userStatus;
-    private int reserve;
+    private long reserve;
     private Boolean emailAgree;
 
     // 주소 정보
